+++ src/frontend/apps/e2e/__tests__/app-impress/doc-export.spec.ts
@@ -7,6 +7,7 @@ import { PDFParse } from 'pdf-parse';
 
 import {
   TestLanguage,
+  clickInEditorMenu,
   createDoc,
   verifyDocName,
   waitForLanguageSwitch,
@@ -24,11 +25,7 @@ test.describe('Doc Export', () => {
     browserName,
   }) => {
     await createDoc(page, 'doc-editor', browserName, 1);
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     await expect(page.getByTestId('modal-export-title')).toBeVisible();
     await expect(
@@ -54,11 +51,7 @@ test.describe('Doc Export', () => {
   test('it exports the doc to docx', async ({ page, browserName }) => {
     const randomDoc = await overrideDocContent({ page, browserName });
 
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     await page.getByRole('combobox', { name: 'Format' }).click();
     await page.getByRole('option', { name: 'Docx' }).click();
@@ -84,11 +77,7 @@ test.describe('Doc Export', () => {
   test('it exports the doc to odt', async ({ page, browserName }) => {
     const randomDoc = await overrideDocContent({ page, browserName });
 
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     await page.getByRole('combobox', { name: 'Format' }).click();
     await page.getByRole('option', { name: 'Odt' }).click();
@@ -139,11 +128,7 @@ test.describe('Doc Export', () => {
     // Give some time for the image to be fully processed
     await page.waitForTimeout(1000);
 
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     await page.getByRole('combobox', { name: 'Format' }).click();
     await page.getByRole('option', { name: 'HTML' }).click();
@@ -231,11 +216,7 @@ test.describe('Doc Export', () => {
       .fill('https://docs.numerique.gouv.fr/assets/logo-gouv.png');
     await page.getByText('Embed image').click();
 
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     await new Promise((resolve) => setTimeout(resolve, 1000));
 
@@ -322,11 +303,7 @@ test.describe('Doc Export', () => {
 
     await overrideDocContent({ page, browserName });
 
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     await page.getByRole('combobox', { name: 'Format' }).click();
     await page.getByRole('option', { name: 'Print' }).click();
@@ -375,11 +352,7 @@ test.describe('Doc Export', () => {
 
     const randomDoc = await overrideDocContent({ page, browserName });
 
-    await page
-      .getByRole('button', {
-        name: 'Export the document',
-      })
-      .click();
+    await clickInEditorMenu(page, 'Download');
 
     const downloadPromise = page.waitForEvent('download', (download) => {
       return download.suggestedFilename().includes(`${randomDoc}.pdf`);