+++ src/frontend/apps/impress/src/features/right-panel/components/RightPanelCollapseButton.tsx
@@ -1,7 +1,5 @@
 import { useEffect, useState } from 'react';
-import { css } from 'styled-components';
 
-import { Card } from '@/components';
 import { CommentSideBarButton } from '@/features/docs/doc-editor/components/comments/CommentSideBar';
 import { useEditorStore } from '@/features/docs/doc-editor/stores/useEditorStore';
 import { useHeadingStore } from '@/features/docs/doc-editor/stores/useHeadingStore';
@@ -30,19 +28,9 @@ export const RightPanelCollapseButton = () => {
   }
 
   return (
-    <Card
-      className="--docs--right-panel-collapse-button"
-      $direction="row"
-      $css={css`
-        padding: var(--c--globals--spacings--xxxs);
-        align-items: center;
-        gap: var(--c--globals--spacings--xxxs);
-        border-radius: var(--c--globals--spacings--xs);
-        box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.05);
-      `}
-    >
+    <>
       {hasHeadings && <TableContentSideBarButton />}
       {hasThreads && <CommentSideBarButton />}
-    </Card>
+    </>
   );
 };